Refactor out common SpectroscopyConfig
