+++ modules/sequence/src/test/scala/lucuma/odb/sequence/gmos/longslit/arb/ArbGmosLongSlitConfig.scala
@@ -27,6 +27,7 @@ import lucuma.core.math.arb.ArbWavelengthDither
 import lucuma.core.model.ExposureTimeMode
 import lucuma.core.model.arb.ArbExposureTimeMode
 import lucuma.core.util.arb.ArbEnumerated
+import lucuma.odb.sequence.gmos.SpectroscopyConfig
 import org.scalacheck.*
 import org.scalacheck.Arbitrary.arbitrary
 
@@ -58,7 +59,7 @@ object ArbGmosLongSlitConfig:
         er <- arbitrary[Option[GmosLongSlitAcquisitionRoi]]
       yield AcquisitionConfig.GmosSouth(e, df, ef, dr, er)
 
-  given Arbitrary[Config.Common] =
+  given Arbitrary[SpectroscopyConfig.Common] =
     Arbitrary:
       for
         w  <- arbitrary[Wavelength]
@@ -72,7 +73,7 @@ object ArbGmosLongSlitConfig:
         r  <- arbitrary[Option[GmosRoi]]
         d  <- arbitrary[Option[List[WavelengthDither]]]
         s  <- arbitrary[Option[List[Q]]]
-      yield Config.Common(
+      yield SpectroscopyConfig.Common(
         w,
         e,
         dx,
@@ -92,7 +93,7 @@ object ArbGmosLongSlitConfig:
         g <- arbitrary[GmosNorthGrating]
         f <- arbitrary[Option[GmosNorthFilter]]
         u <- arbitrary[GmosNorthFpu]
-        c <- arbitrary[Config.Common]
+        c <- arbitrary[SpectroscopyConfig.Common]
         a <- arbitrary[AcquisitionConfig.GmosNorth]
       yield Config.GmosNorth(g, f, u, c, a)
 
@@ -102,6 +103,6 @@ object ArbGmosLongSlitConfig:
         g <- arbitrary[GmosSouthGrating]
         f <- arbitrary[Option[GmosSouthFilter]]
         u <- arbitrary[GmosSouthFpu]
-        c <- arbitrary[Config.Common]
+        c <- arbitrary[SpectroscopyConfig.Common]
         a <- arbitrary[AcquisitionConfig.GmosSouth]
       yield Config.GmosSouth(g, f, u, c, a)
